+++ controllers/AdminMessageController.php
@@ -30,8 +30,6 @@ use skeeks\cms\modules\admin\filters\AdminAccessControl;
 use skeeks\cms\modules\admin\widgets\ControllerActions;
 use skeeks\cms\modules\admin\widgets\DropdownControllerActions;
 use skeeks\cms\rbac\CmsManager;
-use skeeks\cms\validators\db\IsSame;
-use skeeks\cms\validators\HasBehavior;
 use Yii;
 use skeeks\cms\models\User;
 use skeeks\cms\models\searchs\User as UserSearch;