+++ controllers/AdminTreeController.php
@@ -25,8 +25,6 @@ use skeeks\cms\modules\admin\controllers\AdminModelEditorController;
 use skeeks\cms\modules\admin\controllers\helpers\rules\HasModel;
 use skeeks\cms\modules\admin\widgets\ControllerActions;
 use skeeks\cms\modules\admin\widgets\DropdownControllerActions;
-use skeeks\cms\validators\db\IsSame;
-use skeeks\cms\validators\HasBehavior;
 use Yii;
 use skeeks\cms\models\User;
 use skeeks\cms\models\searchs\User as UserSearch;