+++ modules/admin/actions/modelEditor/AdminMultiModelEditAction.php
@@ -17,7 +17,6 @@ use skeeks\cms\modules\admin\filters\AdminAccessControl;
 use skeeks\cms\modules\admin\widgets\ControllerActions;
 use skeeks\cms\modules\admin\widgets\GridViewStandart;
 use skeeks\cms\rbac\CmsManager;
-use skeeks\cms\validators\HasBehavior;
 use skeeks\sx\validate\Validate;
 use yii\authclient\AuthAction;
 use yii\base\View;