+++ modules/admin/widgets/formInputs/SelectModelDialogInput.php
@@ -11,7 +11,6 @@ use skeeks\cms\Exception;
 use skeeks\cms\models\CmsContentElement;
 use skeeks\cms\models\Publication;
 use skeeks\cms\modules\admin\Module;
-use skeeks\cms\validators\HasBehavior;
 use skeeks\sx\validate\Validate;
 use yii\db\ActiveRecord;
 use yii\helpers\Html;