+++ widgets/formInputs/ckeditor/Ckeditor.php
@@ -9,7 +9,6 @@ namespace skeeks\cms\widgets\formInputs\ckeditor;
 
 use skeeks\cms\Exception;
 use skeeks\cms\helpers\UrlHelper;
-use skeeks\cms\validators\HasBehavior;
 use skeeks\sx\validate\Validate;
 use skeeks\yii2\ckeditor\CKEditorWidget;
 use yii\db\ActiveRecord;